[Fix #134673091] Remove old UI tests
