+++ codango/static/js/tests/contact-test.jsx
@@ -1,12 +1,19 @@
 import React from 'react';
-const expect = require('expect');
-import { shallow , mount, render } from 'enzyme';
-// Contact page-test.jsx
+import { expect } from 'chai';
+import { shallow } from 'enzyme';
+import { FormControl } from 'react-bootstrap';
 import Contact from '../components/contact.jsx';
 
-describe('<Contact/>', () => {
-    it("it tests the contact page elements", () => {
-        const wrapper = shallow(<Contact/>);
-        expect(wrapper.find('.row').length).toBe(2);
-    });
+describe('<Contact />', () => {
+  const wrapper = shallow(<Contact />);
+  it('contains Get In touch message', () => {
+    expect(wrapper.contains(<h3>Get In Touch</h3>)).to.equal(true);
+  });
+
+  it('renders contact us form', () => {
+    expect(wrapper.find(FormControl).at(0).prop('name')).to.equal('name');
+    expect(wrapper.find(FormControl).at(1).prop('name')).to.equal('email');
+    expect(wrapper.find(FormControl).at(2).prop('name')).to.equal('subject');
+    expect(wrapper.find(FormControl).at(3).prop('name')).to.equal('message');
+  });
 });
